% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_build.R
\name{sce_build}
\alias{sce_build}
\title{Build sce object from prepared inputs}
\usage{
sce_build(inputs, plotMDS = TRUE)
}
\arguments{
\item{inputs}{Inputs as prepared by calling sce_prepare_inputs AND THEN
EDITING THE PANEL.}

\item{plotMDS}{Boolean specifying whether an MDS plot should be generated for
this sce.}
}
\value{
an SCE object ready to go into the CATALYST pipeline.
}
\description{
This is just a wrapper around CATALYST::prepData that includes defaults for
flow cytometry data and allows you to pass a single inputs object to quickly
build the object without needing to look up all the extra args. Also plots
the MDS by default so you can quickly check if there are sample problems.
}
