% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_volcano.R
\name{sce_volcano}
\alias{sce_volcano}
\title{Draw a differential state volcano plot}
\usage{
sce_volcano(diffState, cutoff, coords = NULL, filter_cluster = NULL)
}
\arguments{
\item{diffState}{The formatted differential state tibble produced by sce_diff_state}

\item{cutoff}{the significance cutoff (using p_adj) you want to use for labeling}

\item{coords}{a named list givin xlim and ylim passed to ggplot}

\item{filter_cluster}{a character vector naming specific cluster_ids that you
want removed from the sce_diff_state results prior to plotting (usually
"unknown" samples that likely represent debris or dead cells).}
}
\value{
a ggplot object for each cluster used in sce_diff_state.
}
\description{
sce_volcano draws a separate volcano plot for every cluster specified in the
annotated clusters of the sce object after running sce_diff_state.
}
\examples{

# Since most sce_diff_state objects will be a double-nested list (the first
# level containing the different sce objects and the second containing the
# different comparisons), you will likely want to call this one with two levels
# of map, like so:

\dontrun{
# don't run automatically, needs a massive sce which is too big to provide in
this package.
  map(diffStates,
      ~map(..1,
           sce_volcano,
           cutoff = 0.05,
           coords = list(xlim = c(-2, 2), ylim = c(0, 3))))
}

}
